Refactor: dto 패키지를 생성하고 엔티티에서 dto로 바꾸는 로직을 dto에 추가한다.
