--- src/main/java/org/prgms/springbootbasic/domain/voucher/VoucherResponseDto.java
@@ -1,10 +1,0 @@
-package org.prgms.springbootbasic.domain.voucher;
-
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public record VoucherResponseDto(UUID voucherId,
-                                 long discountDegree,
-                                 String voucherPolicy,
-                                 LocalDateTime createdAt) {
-}
