+++ src/main/java/org/prgms/springbootbasic/domain/voucher/dto/VoucherRequestDto.java
@@ -1,4 +1,4 @@
-package org.prgms.springbootbasic.domain.voucher;
+package org.prgms.springbootbasic.domain.voucher.dto;
 
 public record VoucherRequestDto (String voucherId, String voucherPolicy, long discountDegree) {
 }
