+++ src/test/java/org/prgms/springbootbasic/controller/voucher/VoucherApiControllerTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.prgms.springbootbasic.domain.VoucherType;
 import org.prgms.springbootbasic.domain.voucher.Voucher;
-import org.prgms.springbootbasic.domain.voucher.VoucherRequestDto;
+import org.prgms.springbootbasic.domain.voucher.dto.VoucherRequestDto;
 import org.prgms.springbootbasic.service.VoucherService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
